+++ backend/globals.js
@@ -1,1 +1,5 @@
+// -- Excalidraw settings --
+// @TODO replace with "true" once Preact is integrated to ExcalidrawCanvas
 var process = { env: { IS_PREACT: "false" } };
+window.__REACT_DEVTOOLS_GLOBAL_HOOK__ = { isDisabled: true };
+// -- / --
